SC-proyect/fraud-detection-system/requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# NLP
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0

# Geospatial
geopy>=2.3.0
folium>=0.14.0
haversine>=2.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# API & Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# UI
streamlit>=1.25.0
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Flask-CORS==4.0.0 

# Data Validation
great-expectations>=0.17.0
pandera>=0.16.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Logging & Monitoring
loguru>=0.7.0

# Kaggle Dataset Download
kaggle>=1.5.16

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
nbconvert>=7.7.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
pylint>=2.17.0
mypy>=1.5.0
